FLA-1873 Move Save File to FileUtil
